// src/app/shared/enums/issue-status.enum.ts
export enum IssueStatusEnum {
  NOVA = 1,
  EM_ANDAMENTO = 2,
  EM_TESTES = 3,
  ERRO_NO_TESTE = 4,
  FECHADA = 5,
  NAO_AUTORIZADA = 6,
  REQUISICAO = 7,
  HOMOLOGADA = 8,
  EM_PAUSA = 9,
  COMERCIAL = 10,
  TESTES_EM_ANDAMENTO = 11,
  TESTES_EM_PAUSA = 12,
  EM_TESTES_CLIENTE = 13,
  ATUALIZACAO_DISPONIVEL = 14,
  AJUSTE_NA_ANALISE = 15,
  NAO_APROVADA_COMERCIAL = 16,
  APROVADA_COMERCIAL = 17,
  ENVIADO_PROPOSTA = 18,
  TRIAGEM = 19,
  ANALISADA_PARA_COMERCIAL = 21,
  EM_DOCUMENTACAO = 23,
  EM_REVISAO = 24,
  BACKLOG = 25,
  RETORNO_CODE_REVIEW = 26,
  AGUARDANDO_AGENDAMENTO = 27,
  AGUARDANDO_DOCUMENTACAO = 28
}